rearrange and cleanup
