+++ KVMarketSquare/UIComponents/MapView.swift
@@ -1,0 +1,38 @@
+//
+//  MapView.swift
+//  KVMarketSquare
+//
+//  Created by Kevin Taniguchi on 12/16/21.
+//
+
+import Foundation
+import MapKit
+import SwiftUI
+
+struct MapView : View {
+    var locationResult : MKLocalSearchCompletion
+    @StateObject private var viewModel = AddressToCoordinateResolver()
+    
+    struct Marker: Identifiable {
+        let id = UUID()
+        var location: MapMarker
+    }
+    
+    var body: some View {
+        Group {
+            if viewModel.isLoading {
+                Text("Loading...")
+            } else {
+                Map(coordinateRegion: $viewModel.region,
+                    annotationItems: [Marker(location: MapMarker(coordinate: viewModel.coordinateForMap))]) { (marker) in
+                    marker.location
+                }
+            }
+        }.onAppear {
+            viewModel.reconcileLocation(location: locationResult)
+        }.onDisappear {
+            viewModel.clear()
+        }
+        .navigationTitle(Text(locationResult.title))
+    }
+}
